Remove FullRoomException, a few debug messages and more
FullRoomException was never managed. If a room is full, it is managed without this exception.
The debug messages remove were used to understand how disconnections worked and are no longer needed.
Connection settings loading has been remove: it wasn't used and the hostname can be set from the command line

Signed-off-by: Matteo Delton <deltonmatteo@gmail.com>